requirements.txt
anyio==3.6.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==4.4.2
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
Pillow==9.3.0
proglog==0.1.10
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
redis==4.5.1
starlette==0.22.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
websockets==10.4